+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add retry on (gitlab) rateLimit ([#30](https://github.com/diffplug/blowdryer/pull/30))
 
 ## [1.5.1] - 2022-01-31
 ### Fixed